# protoc-play/api/books.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "books.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/books": {
      "get": {
        "operationId": "ListBooks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoc_playListBooksResponse"
            }
          }
        },
        "tags": [
          "BookService"
        ]
      },
      "post": {
        "operationId": "CreateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoc_playBook"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoc_playBook"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/books/{book.book_id}": {
      "patch": {
        "operationId": "UpdateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoc_playBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book.book_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoc_playBook"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/books/{book_id}": {
      "get": {
        "operationId": "GetBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoc_playBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      },
      "delete": {
        "operationId": "DeleteBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    }
  },
  "definitions": {
    "protoc_playBook": {
      "type": "object",
      "properties": {
        "book_id": {
          "type": "string"
        }
      }
    },
    "protoc_playListBooksResponse": {
      "type": "object",
      "properties": {
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoc_playBook"
          }
        }
      }
    }
  }
}